@@ -1,16 +1,16 @@
 {
   "name": "kalm",
-  "version": "7.0.0",
+  "version": "8.0.0",
   "description": "The socket optimizer",
-  "main": "bin/kalm.js",
+  "main": "dist/kalm.js",
   "scripts": {
-    "build": "../../scripts/build.sh kalm",
-    "clean": "../../scripts/cleanup.sh",
+    "build": "npm run clean && rollup -c rollup.config.mjs",
+    "clean": "rm -rf ./dist/*",
     "test": "jest ./tests",
-    "prepublish": "cp ../../LICENSE ./LICENSE && cp ../../CHANGELOG.md ./CHANGELOG.md"
+    "prepublishOnly": "npm run build && cp ../../LICENSE ./LICENSE && cp ../../types.d.ts ./types.d.ts"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=20"
   },
   "funding": {
     "type": "Open Collective",
@@ -34,7 +34,7 @@
     "peer"
   ],
   "files": [
-    "bin",
+    "dist",
     "types.d.ts"
   ],
   "typings": "./types.d.ts",
@@ -53,8 +53,7 @@
     "transform": {
       "^.+\\.tsx?$": [
       "ts-jest", {
-        "diagnostics": false,
-        "isolatedModules": true
+        "diagnostics": false
       }]
     }
   }
