@@ -1,16 +1,16 @@
 {
   "name": "@kalm/tcp",
-  "version": "7.0.0",
+  "version": "8.0.0",
   "description": "TCP transport for Kalm",
-  "main": "bin/tcp.js",
+  "main": "dist/tcp.js",
   "scripts": {
-    "build": "../../scripts/build.sh tcp",
-    "clean": "../../scripts/cleanup.sh",
+    "build": "npm run clean && rollup -c rollup.config.mjs",
+    "clean": "rm -rf ./dist/*",
     "test": "jest ./tests",
-    "prepublish": "cp ../../LICENSE ./LICENSE && cp ../../CHANGELOG.md ./CHANGELOG.md"
+    "prepublishOnly": "npm run build && cp ../../LICENSE ./LICENSE"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=20"
   },
   "funding": {
     "type": "Open Collective",
@@ -28,7 +28,7 @@
     "nagle",
     "tcp"
   ],
-  "files": ["bin", "types.d.ts"],
+  "files": ["dist", "types.d.ts"],
   "typings": "./types.d.ts",
   "author": "frederic charette <fredericcharette@gmail.com>",
   "license": "Apache-2.0",
@@ -48,8 +48,7 @@
     "transform": {
       "^.+\\.tsx?$": [
       "ts-jest", {
-        "diagnostics": false,
-        "isolatedModules": true
+        "diagnostics": false
       }]
     }
   }
