@@ -1,15 +1,5 @@
-/**
- * KALM Benchmark
- */
-
-
-/* Requires ------------------------------------------------------------------*/
-
-const dgram = require('dgram');
-
-const settings = require('../settings');
-
-/* Local letiables -----------------------------------------------------------*/
+import { createSocket } from 'dgram';
+import settings from '../settings.ts';
 
 let server;
 let client;
@@ -17,14 +7,13 @@ let client;
 let count = 0;
 let handbreak = true;
 
-/* Methods -------------------------------------------------------------------*/
-
 function _absorb(err) {
-  console.log(err); /* eslint-disable-line */
+  console.log(err);
+  return true;
 }
 
-function setup(resolve) {
-  server = dgram.createSocket('udp4');
+export function setup(resolve) {
+  server = createSocket('udp4');
   server.on('message', () => {
     server.send(Buffer.from(JSON.stringify(settings.testPayload)), 1111, '0.0.0.0');
   });
@@ -34,7 +23,7 @@ function setup(resolve) {
   resolve();
 }
 
-function teardown(resolve) {
+export function teardown(resolve) {
   server.close(() => {
     server = null;
     client = null;
@@ -42,15 +31,15 @@ function teardown(resolve) {
   });
 }
 
-function stop(resolve) {
+export function stop(resolve) {
   handbreak = true;
   setTimeout(resolve, 0);
 }
 
-function step(resolve) {
+export function step(resolve) {
   if (handbreak) return;
   if (!client) {
-    client = dgram.createSocket('udp4');
+    client = createSocket('udp4');
     client.on('error', _absorb);
     client.on('message', () => count++);
     client.bind(1111, '0.0.0.0');
@@ -59,12 +48,3 @@ function step(resolve) {
   client.send(Buffer.from(JSON.stringify(settings.testPayload)), settings.port, '0.0.0.0');
   resolve();
 }
-
-/* Exports -------------------------------------------------------------------*/
-
-module.exports = {
-  setup,
-  teardown,
-  step,
-  stop,
-};
