@@ -1,4 +1,4 @@
-import net from 'net';
+import net from 'node:net';
 
 interface IPCSocket extends net.Socket {
   server?: {
@@ -15,9 +15,9 @@ interface IPCSocket extends net.Socket {
 type IPCConfig = {
   socketTimeout?: number
   path?: string
-}
+};
 
-function ipc({ socketTimeout = 30000, path = '/tmp/app.socket-' }: IPCConfig = {}): KalmTransport {
+export default function ipc({ socketTimeout = 30000, path = '/tmp/app.socket-' }: IPCConfig = {}): KalmTransport {
   return function socket(params: ClientConfig, emitter: NodeJS.EventEmitter): Socket {
     let listener: net.Server;
 
@@ -44,23 +44,24 @@ function ipc({ socketTimeout = 30000, path = '/tmp/app.socket-' }: IPCConfig = {
     function connect(handle: IPCSocket): IPCSocket {
       const connection: net.Socket = handle || net.connect(`${path}${params.port}`);
       let buffer = '';
-      connection.on('data', req => {
+      connection.on('data', (req) => {
         buffer += req.toString();
         const chunks = buffer.split('\n\n');
         if (buffer.substring(buffer.length - 2) !== '\n\n') {
           buffer = chunks[chunks.length - 1];
           chunks.pop();
-        } else {
+        }
+        else {
           buffer = '';
         }
 
         for (let i = 0; i < chunks.length; i++) {
-          if (chunks[i] !== '') emitter.emit('frame', JSON.parse(chunks[i]), req.length);
+          if (chunks[i] !== '') emitter.emit('frame', { body: JSON.parse(chunks[i]), payloadBytes: req.length });
         }
       });
       connection.on('error', err => emitter.emit('error', err));
-      connection.on('connect', () => emitter.emit('connect', connection));
-      connection.on('close', () => emitter.emit('disconnect'));
+      connection.on('connect', () => emitter.emit('connect'));
+      connection.on('close', () => emitter.emit('disconnected'));
       connection.setTimeout(socketTimeout, () => disconnect(handle));
       return connection;
     }
@@ -83,6 +84,3 @@ function ipc({ socketTimeout = 30000, path = '/tmp/app.socket-' }: IPCConfig = {
     };
   };
 }
-
-// Ensures support for modules and requires
-module.exports = ipc;
