@@ -1,4 +1,4 @@
-import { EventEmitter } from 'events';
+import { EventEmitter } from '../utils/events';
 import { log } from '../utils/logger';
 
 export function Client(params: ClientConfig, emitter: EventEmitter, socket?: any): Client {
@@ -7,14 +7,14 @@ export function Client(params: ClientConfig, emitter: EventEmitter, socket?: any
   type Channel = {
     name: string
     packets: any[]
-    handlers: Function[]
-  }
+    handlers: ((packet: any, context: Context) => void)[]
+  };
 
-  const channels: {[channel: string]: Channel } = {};
+  const channels: { [channel: string]: Channel } = {};
   const routine = params.routine(params, _wrap);
   const transport: Socket = params.transport(params, emitter);
   let instance = null;
-  
+
   const remote: Remote = (params.isServer) ? transport.remote(socket) : { host: params.host, port: params.port };
   const local: Remote = (params.isServer) ? { host: params.host, port: params.port } : null;
 
@@ -35,11 +35,13 @@ export function Client(params: ClientConfig, emitter: EventEmitter, socket?: any
 
   function _wrap(frameId: number): void {
     transport.send(socket, getChannels().reduce((frame, channelName) => {
-      frame.channels[channelName] = channels[channelName].packets;
+      if (channels[channelName].packets.length > 0) frame.channels[channelName] = channels[channelName].packets;
       return frame;
     }, { frameId, channels: {} }));
 
-    getChannels().forEach(channelName => { channels[channelName].packets.length = 0; });
+    for (const channelName in channels) {
+      channels[channelName].packets.length = 0;
+    }
   }
 
   function _handleConnect(): void {
@@ -51,9 +53,10 @@ export function Client(params: ClientConfig, emitter: EventEmitter, socket?: any
     log(`error ${err.message}`);
   }
 
-  function _handleRequest(frame: RawFrame, payloadBytes: number): void {
+  function _handleRequest(frameEvent: { body: RawFrame, payloadBytes: number }): void {
+    const frame = frameEvent.body;
     if (frame && frame.channels) {
-      Object.keys(frame.channels).forEach(channelName => {
+      Object.keys(frame.channels).forEach((channelName) => {
         frame.channels[channelName].forEach((packet, messageIndex) => {
           if (channelName in channels) {
             channels[channelName].handlers.forEach(handler => handler(
@@ -64,7 +67,7 @@ export function Client(params: ClientConfig, emitter: EventEmitter, socket?: any
                   channel: channelName,
                   id: frame.frameId,
                   messageIndex,
-                  payloadBytes,
+                  payloadBytes: frameEvent.payloadBytes,
                   payloadMessages: frame.channels[channelName].length,
                 },
               },
@@ -78,6 +81,8 @@ export function Client(params: ClientConfig, emitter: EventEmitter, socket?: any
   function _handleDisconnect() {
     connected = 0;
     log(`lost connection to ${params.host}:${params.port}`);
+    // Gives a change for other internal listeners to run their business logic
+    setTimeout(() => emitter.emit('disconnect'), 0);
   }
 
   function write(channelName: string, message: Serializable): void {
@@ -93,22 +98,23 @@ export function Client(params: ClientConfig, emitter: EventEmitter, socket?: any
     if (connected > 1) setTimeout(() => transport.disconnect(socket), 0);
   }
 
-  function subscribe(channel: string, handler: (msg: any, frame: Frame) => void): void {
-    _resolveChannel(channel).handlers.push(handler);
+  function subscribe(channelName: string, handler: (msg: any, context: Context) => void): void {
+    _resolveChannel(channelName).handlers.push(handler);
   }
 
-  function unsubscribe(channelName: string, handler?: (msg: any, frame: Frame) => void): void {
+  function unsubscribe(channelName: string, handler?: (msg: any, context: Context) => void): void {
     if (!(channelName in channels)) return;
     if (handler) {
       const index = channels[channelName].handlers.indexOf(handler);
       if (index > -1) channels[channelName].handlers.splice(index, 1);
-    } else channels[channelName].handlers = [];
+    }
+    else channels[channelName].handlers = [];
 
     if (channels[channelName].handlers.length === 0 && channels[channelName].packets.length === 0) delete channels[channelName];
   }
 
   emitter.on('connect', _handleConnect);
-  emitter.on('disconnect', _handleDisconnect);
+  emitter.on('disconnected', _handleDisconnect);
   emitter.on('error', _handleError);
   emitter.on('frame', _handleRequest);
   if (!socket) log(`connecting to ${params.host}:${params.port}`);
