@@ -1,5 +1,5 @@
 import { Client } from '../../../src/components/client';
 
 describe('Client', () => {
-  it('TODO', () => { expect(Client).not.toBeUndefined(); });
+  it('TODO', () => expect(Client).not.toBeUndefined());
 });
