@@ -1,5 +1,5 @@
 import { Server } from '../../../src/components/server';
 
 describe('Provider', () => {
-  it('TODO', () => { expect(Server).not.toBeUndefined(); });
+  it('TODO', () => expect(Server).not.toBeUndefined());
 });
