@@ -1,5 +1,5 @@
 import { dynamic } from '../../../src/routines/dynamic';
 
 describe('Dynamic routine', () => {
-  it('TODO', () => { expect(dynamic).not.toBeUndefined(); });
+  it('TODO', () => expect(dynamic).not.toBeUndefined());
 });
