@@ -1,5 +1,5 @@
 import { realtime } from '../../../src/routines/realtime';
 
 describe('Realtime routine', () => {
-  it('TODO', () => { expect(realtime).not.toBeUndefined(); });
+  it('TODO', () => expect(realtime).not.toBeUndefined());
 });
