@@ -1,5 +1,5 @@
 import { tick } from '../../../src/routines/tick';
 
 describe('Tick routine', () => {
-  it('TODO', () => { expect(tick).not.toBeUndefined(); });
+  it('TODO', () => expect(tick).not.toBeUndefined());
 });
