@@ -1,5 +1,5 @@
 import { log } from '../../../src/utils/logger';
 
 describe('Logger util', () => {
-  it('TODO', () => { expect(log).not.toBeUndefined(); });
+  it('TODO', () => expect(log).not.toBeUndefined());
 });
