@@ -1,5 +1,4 @@
 import net from 'net';
-import events from 'events';
 
 interface TCPSocket extends net.Socket {
   _peername: {
@@ -12,8 +11,8 @@ interface TCPConfig {
   socketTimeout?: number
 }
 
-function tcp({ socketTimeout = 30000 }: TCPConfig = {}): KalmTransport {
-  return function socket(params: ClientConfig, emitter: events.EventEmitter): Socket {
+export default function tcp({ socketTimeout = 30000 }: TCPConfig = {}): KalmTransport {
+  return function socket(params: ClientConfig, emitter: NodeJS.EventEmitter): Socket {
     let listener: net.Server;
 
     function bind(): void {
@@ -22,7 +21,7 @@ function tcp({ socketTimeout = 30000 }: TCPConfig = {}): KalmTransport {
       listener.listen(params.port, () => emitter.emit('ready'));
     }
 
-    function remote(handle: TCPSocket): Remote {
+    function remote(handle?: TCPSocket): Remote {
       return {
         host: handle?.remoteAddress || handle?._peername?.address || null,
         port: handle?.remotePort || handle?._peername?.port || null,
@@ -39,23 +38,24 @@ function tcp({ socketTimeout = 30000 }: TCPConfig = {}): KalmTransport {
     function connect(handle: TCPSocket): TCPSocket {
       const connection: net.Socket = handle || net.connect(params.port, params.host);
       let buffer = '';
-      connection.on('data', req => {
+      connection.on('data', (req) => {
         buffer += req.toString();
         const chunks = buffer.split('\n\n');
         if (buffer.substring(buffer.length - 2) !== '\n\n') {
           buffer = chunks[chunks.length - 1];
           chunks.pop();
-        } else {
+        }
+        else {
           buffer = '';
         }
 
         for (let i = 0; i < chunks.length; i++) {
-          if (chunks[i] !== '') emitter.emit('frame', JSON.parse(chunks[i]), req.length);
+          if (chunks[i] !== '') emitter.emit('frame', { body: JSON.parse(chunks[i]), payloadBytes: req.length });
         }
       });
       connection.on('error', err => emitter.emit('error', err));
-      connection.on('connect', () => emitter.emit('connect', connection));
-      connection.on('close', () => emitter.emit('disconnect'));
+      connection.on('connect', () => emitter.emit('connect'));
+      connection.on('close', () => emitter.emit('disconnected'));
       connection.setTimeout(socketTimeout, () => disconnect(handle));
       return connection as TCPSocket;
     }
@@ -78,6 +78,3 @@ function tcp({ socketTimeout = 30000 }: TCPConfig = {}): KalmTransport {
     };
   };
 }
-
-// Ensures support for modules and requires
-module.exports = tcp;
