@@ -1,42 +1,42 @@
-import events from 'events';
-import {createServer} from 'https';
+import { WebSocket as WSClient, WebSocketServer } from 'ws';
 
-const isBrowser = (typeof WebSocket !== 'undefined');
-const WS = isBrowser ? WebSocket : require('ws');
+const nativeAPIExists = (typeof WebSocket !== 'undefined');
 
 type WSConfig = {
   cert?: string
   key?: string
-  agent?: any
   socketTimeout?: number
-}
+};
 
 type WSHandle = WebSocket & {
-  _queue: string[],
-  _timer: ReturnType<typeof setTimeout>
+  _queue?: string[]
+  _timer?: ReturnType<typeof setTimeout>
   headers?: any
   connection?: any
   _socket?: any
-}
+};
 
-
-function ws({ cert, key, agent, socketTimeout = 30000 }: WSConfig = {}): KalmTransport {
-  return function socket(params: ClientConfig, emitter: events.EventEmitter): Socket {
+export default function ws({ cert, key, socketTimeout = 30000 }: WSConfig = {}): KalmTransport {
+  return function socket(params: ClientConfig, emitter: NodeJS.EventEmitter): Socket {
     let listener;
 
-    function bind(): void {
+    async function bind(): Promise<void> {
+      if (typeof window !== 'undefined') throw new Error('Cannot create a websocket server from the browser');
+
       if (cert && key) {
-        const server = createServer({ key, cert }, req => req.socket.end());
-        listener = new WS.Server({ port: params.port, server });
-      } else {
-        listener = new WS.Server({ port: params.port });
+        const https = await import('https');
+        const server = https.createServer({ key, cert }, req => req.socket.end());
+        listener = new WebSocketServer({ port: params.port, server });
+      }
+      else {
+        listener = new WebSocketServer({ port: params.port });
       }
       listener.on('connection', soc => emitter.emit('socket', soc));
       listener.on('error', err => emitter.emit('error', err));
-      emitter.emit('ready');
+      setTimeout(() => emitter.emit('ready'), 1);
     }
 
-    function send(handle: WSHandle & { _queue: string[] }, payload: RawFrame | string): void {
+    function send(handle: WSHandle & { _queue?: string[] }, payload: RawFrame | string): void {
       if (handle && handle.readyState === 1) handle.send(typeof payload === 'string' ? payload : JSON.stringify(payload));
       else handle._queue.push(JSON.stringify(payload));
       resetTimeout(handle);
@@ -48,19 +48,19 @@ function ws({ cert, key, agent, socketTimeout = 30000 }: WSConfig = {}): KalmTra
 
     function connect(handle?: WSHandle): WSHandle {
       const protocol: string = (!!cert && !!key) === true ? 'wss' : 'ws';
-      const connection: WSHandle = handle || new WS(`${protocol}://${params.host}:${params.port}`, { ...(agent ? {agent} : {})});
+      const connection: WSHandle = handle || new (nativeAPIExists ? WebSocket : WSClient)(`${protocol}://${params.host}:${params.port}`);
       connection.binaryType = 'arraybuffer';
-      const evtType: string = isBrowser ? 'addEventListener' : 'on';
+      const evtType: string = nativeAPIExists ? 'addEventListener' : 'on';
       connection._queue = [];
       connection._timer = null;
-      connection[evtType]('message', evt => {
-        emitter.emit('frame', JSON.parse(evt.data || evt), (evt.data || evt).length);
+      connection[evtType]('message', (evt) => {
+        emitter.emit('frame', { body: JSON.parse(evt.data || evt), payloadBytes: (evt.data || evt).length });
         resetTimeout(connection);
       });
       connection[evtType]('error', err => emitter.emit('error', err));
-      connection[evtType]('close', () => emitter.emit('disconnect'));
+      connection[evtType]('close', () => emitter.emit('disconnected'));
       connection[evtType]('open', () => {
-        emitter.emit('connect', connection);
+        emitter.emit('connect');
         connection._queue.forEach(payload => send(connection, payload));
 
         resetTimeout(connection);
@@ -77,7 +77,7 @@ function ws({ cert, key, agent, socketTimeout = 30000 }: WSConfig = {}): KalmTra
     }
 
     function remote(handle: WSHandle): Remote {
-      const h = handle && handle.headers || {};
+      const h = (handle && handle.headers) || {};
       const headerHost = h && h['x-forwarded-for'] && h['x-forwarded-for'].split(',')[0];
       const remoteHost = handle?.connection?.remoteAddress;
       const socketHost = handle?._socket?.server?._connectionKey;
@@ -90,7 +90,7 @@ function ws({ cert, key, agent, socketTimeout = 30000 }: WSConfig = {}): KalmTra
     function disconnect(handle) {
       if (handle) {
         clearTimeout(handle._timer);
-        handle.end();
+        if (handle.end) handle.end();
         handle.close();
       }
     }
@@ -105,6 +105,3 @@ function ws({ cert, key, agent, socketTimeout = 30000 }: WSConfig = {}): KalmTra
     };
   };
 }
-
-// Ensures support for modules and requires
-module.exports = ws;
